+++ packages/web-ui/src/Book/Author/styles.ts
@@ -1,7 +1,7 @@
 import { merge } from 'lodash';
 import { lineClamp, resetFont, resetLayout } from '../../styles';
 
-export const author = merge({}, resetFont, resetLayout, lineClamp(2), {
+export const author = merge({}, resetFont, resetLayout, lineClamp(1), {
   lineHeight: '1.4em',
   marginTop: 4,
   color: '#808991',