+++ packages/web-ui/src/Book/PortraitBook/styles.ts
@@ -1,15 +1,16 @@
 import { PositionProperty } from "csstype";
+import { THUMBNAIL_HEIGHT_RATIO } from "../ThumbnailImage/styles";
 
-export const portraitBook = {
-  maxWidth: 100,
-};
+export const portraitBook = (thumbnailWidth: number | string) => ({
+  width: thumbnailWidth,
+});
 
 export const thumbnail = {
   position: 'relative' as PositionProperty,
   lineHeight: 0,
   width: '100%',
   height: 'auto',
-  paddingBottom: '162%', // width : height = 1: 1.618
+  paddingBottom: THUMBNAIL_HEIGHT_RATIO,
   '& .Thumbnail': {
     position: 'absolute' as PositionProperty,
     bottom: 0,