+++ packages/web-ui/src/Book/Thumbnail/index.tsx
@@ -42,14 +42,45 @@ export interface ThumbnailProps extends
     selected?: boolean;
     readingStatus?: ReadingStatus;
     ridiSelect?: boolean;
+    thumbnailChildrenSize?: ThumbnailChildrenSize;
     thumbnailLink?: React.ReactElement<any>;
-    thumbnailWidth?: number;
     unitBook?: boolean;
     unitBookCount?: React.ReactElement<UnitBookCountProps>;
     updateBadge?: boolean;
     viewType?: ViewType;
     [extraKey: string]: any;
   }
+
+export enum ThumbnailChildrenSize {
+  XXSmall = 'XXSmall',
+  XSmall = 'XSmall',
+  Small = 'Small',
+  Medium = 'Medium',
+  Large = 'Large',
+  XLarge = 'XLarge',
+  XXLarge = 'XXLarge',
+}
+
+const getThumbnailChildrenSize = (width: string | number) => {
+  const thumbnailWidthString = String(width).split('px')[0];
+  const thumbnailWidth = Number(thumbnailWidthString);
+  if (thumbnailWidthString.indexOf('%') !== -1) {
+    // % width default size
+    return ThumbnailChildrenSize.Medium;
+  } else if (thumbnailWidth >= 180) {
+    return ThumbnailChildrenSize.XLarge;
+  } else if (thumbnailWidth >= 130) {
+    return ThumbnailChildrenSize.Large;
+  } else if (thumbnailWidth >= 110) {
+    return ThumbnailChildrenSize.Medium;
+  } else if (thumbnailWidth >= 100) {
+    return ThumbnailChildrenSize.Small;
+  } else if (thumbnailWidth >= 80) {
+    return ThumbnailChildrenSize.XSmall;
+  } else {
+    return ThumbnailChildrenSize.XXSmall;
+  }
+};
 
 export const Thumbnail: React.FunctionComponent<ThumbnailProps> = (props) => {
   const {
@@ -67,6 +98,7 @@ export const Thumbnail: React.FunctionComponent<ThumbnailProps> = (props) => {
     ridiselect,
     selected = false,
     selectMode = false,
+    thumbnailChildrenSize,
     thumbnailLink,
     thumbnailTitle,
     thumbnailUrl,
@@ -77,6 +109,7 @@ export const Thumbnail: React.FunctionComponent<ThumbnailProps> = (props) => {
     viewType = ViewType.Portrait,
     ...extraProps
   } = props;
+  const childrenSize = thumbnailChildrenSize ? thumbnailChildrenSize : thumbnailWidth ? getThumbnailChildrenSize(thumbnailWidth) : ThumbnailChildrenSize.Medium;
 
   return (
     <div
@@ -96,16 +129,17 @@ export const Thumbnail: React.FunctionComponent<ThumbnailProps> = (props) => {
           <ThumbnailCheckbox
             onChange={onSelectedChange}
             checked={selected}
+            size={childrenSize}
           />
         }
-        <ThumbnailImage thumbnailUrl={thumbnailUrl} thumbnailTitle={thumbnailTitle} />
+        <ThumbnailImage thumbnailUrl={thumbnailUrl} thumbnailTitle={thumbnailTitle} thumbnailWidth={thumbnailWidth} />
         {adultBadge && <AdultBadge />}
         {updateBadge && <UpdateBadge />}
+        {(notAvailable || selectMode || expired) && <div css={styles.thumbnailDimmed} />}
         {viewType === ViewType.Portrait &&
           <React.Fragment>
             {unitBook && (
               <React.Fragment>
-                {(notAvailable || selectMode) && <div css={styles.thumbnailDimmed} />}
                 {unitBookCount}
                 {downloadStatus === DownloadStatus.Downloading && !selectMode && <UnitBookDownloading size={UnitBookDownloadingSize.Large} />}
               </React.Fragment>
@@ -119,20 +153,19 @@ export const Thumbnail: React.FunctionComponent<ThumbnailProps> = (props) => {
                     size={DownloadButtonSize.Large}
                   />
                 }
-                {(notAvailable || selectMode) && <div css={styles.thumbnailDimmed} />}
                 {ridiselect ? (
-                  <Ridiselect />
+                  <Ridiselect size={childrenSize} />
                 ) : expired ? (
-                  <Expired />
+                  <Expired size={childrenSize} />
                 ) : expiredAt ? (
-                  <ExpiredAt expiredAt={expiredAt} />
+                  <ExpiredAt expiredAt={expiredAt} size={childrenSize} />
                 ) : null}
               </React.Fragment>
             )}
           </React.Fragment>
         }
         {children}
-        {notAvailable && !selectMode && <NotAvailable />}
+        {notAvailable && !selectMode && <NotAvailable size={childrenSize} />}
         {thumbnailLink && (
           <div css={styles.thumbnailLink}>{thumbnailLink}</div>
         )}