+++ packages/web-ui/src/Book/ThumbnailCheckbox/index.tsx
@@ -1,18 +1,21 @@
 /** @jsx jsx */
 import { jsx } from '@emotion/core';
-import { check5 as CheckIcon } from '@ridi/web-icons';
+import { checkboxCircle as CheckIcon } from '@ridi/web-icons';
 import * as React from 'react';
+import { ThumbnailChildrenSize } from '../Thumbnail';
 import * as styles from './styles';
 
 export interface ThumbnailCheckboxProps {
   checked: boolean;
   onChange: (e: React.SyntheticEvent<any>) => void;
+  size?: ThumbnailChildrenSize;
 }
 
 export const ThumbnailCheckbox: React.FunctionComponent<ThumbnailCheckboxProps> = (props) => {
   const {
     checked,
     onChange,
+    size,
   } = props;
 
   return (
@@ -23,7 +26,7 @@ export const ThumbnailCheckbox: React.FunctionComponent<ThumbnailCheckboxProps> 
         checked={checked}
         onChange={onChange}
       />
-      <span css={styles.iconWrapper}>
+      <span css={styles.iconWrapper(size)}>
         <CheckIcon css={styles.checkboxIcon(checked)} />
       </span>
     </label>