+++ packages/web-ui/src/Book/UpdateBadge/index.tsx
@@ -1,8 +1,10 @@
 /** @jsx jsx */
 import { jsx } from '@emotion/core';
+import { up as UpIcon} from '@ridi/web-icons';
 import * as React from 'react';
+import { displayHidden } from '../../styles';
 import * as styles from './styles';
 
 export const UpdateBadge: React.FunctionComponent = () => (
-  <p css={styles.updateBadge}>UP</p>
+  <p css={styles.updateBadge}><UpIcon css={styles.updateIcon}/><span css={displayHidden}>신규 업데이트 도서</span></p>
 );