+++ CHANGELOG.md
@@ -5,8 +5,69 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.3.10] - 2019-03-18
+### Fixed
+- Fix line clamp bug
+- Fix ExpiredAt remain time conditional rendering
+- Fix typo
+
+## [0.3.9] - 2019-03-14
+### Fixed
+- Fix some method for IE : includes to indexOf
+
+## [0.3.8] - 2019-03-13
+### Fixed
+- Fix wrong font style order
+
+### Added
+- Add conditional rendering to ExpiredAt icon
+- Add thumbnail image skeleton’s height Ratio
+
+## [0.3.7] - 2019-03-13
+### Added
+- Add thumbnailChildrenSize props
+
+### Fixed
+- Fix update badge style
+
+## [0.3.6] - 2019-02-26
+### Added
+- Add a logic for already loaded image
+
+## [0.3.5] - 2019-02-26
+### Added
+- missing props
+
+### Fixed
+- flex layout
+- README
+
+## [0.3.4] - 2019-02-25
+### Fixed
+- Fix flex style for IE 11
+
+## [0.3.3] - 2019-02-21
+### Added
+- Add Thumbnail Image skeleton
+
+### Fixed
+- Fix Expired layout
+
+## [0.3.2] - 2019-02-20
+### Added
+- Fix/book component (#48)
+- Setup Netlify (#47)
+
+## [0.3.1] - 2019-02-15
+### Added
+- Add slots to LandscapeBook and PortraitBook (#46)
+
+## [0.3.0] - 2019-02-14
 ### Added
 - Implement `Group` component. (#35)
+
+### Changed
+- **BREAKING**: Remove unused icons and renaming, Fix Book component styles (#44)
 
 ## [0.2.2] - 2019-02-08
 ### Fixed